requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
vapi_server_sdk
google-genai
python-dotenv
requests
fish-audio-sdk
flask>=3.0.0